requirements.txt
matrix-nio
diplomacy
